requirements.txt
# Core ML & Data Science
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.4.0
joblib==1.3.2
scipy==1.12.0

# NLP & Text Processing
nltk==3.8.1
textblob==0.18.0

# Groq AI Integration
groq==0.4.2

# API & Web
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
python-multipart==0.0.7

# Visualization & Reporting
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
rich==13.7.0
click==8.1.7

# Logging & Monitoring
loguru==0.7.2
